Add files via upload
This is the first task of the internship in java programming titled Number Game where a random number from a specified range 1 to 100 is generated and then user gives the number where it compares whether the number is same ,high or low then the chosen number. Accordingly, a particular number of trials are provided ,thereafter the number is displayed and then the score is displayed with the number of attempts taken.