Add files via upload
This task of the given internship was to provide a marks based calculator for a student. As marks of around 5 subjects are taken then total and average of all the marks is calculated and according to the average scored the grade is rewarded. This task has been made using basic conditional statements.