Add files via upload
This task of the internship was to provide an ATM Interface for the user showing different methods for withdrawing, depositing and checking the balance. It also represents a class which shows user's bank account and the balance in it. Then further options are chosen accordingly which gives the result as the success or failure of the transaction.